% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-sizecomp.R
\name{get_sizecomp}
\alias{get_sizecomp}
\title{Get observed and predicted size composition values}
\usage{
get_sizecomp(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
List of observed and predicted size composition values
}
\description{
TODO: Insert more information here.
}

