% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_catch}
\alias{plot_catch}
\title{Plot observed and predicted catch values}
\usage{
plot_catch(replist, plot_res = FALSE)
}
\arguments{
\item{replist}{List object created by read_admb function}

\item{plot_res}{plot residuals only (default=F)}
}
\value{
Plot of catch history (observed) and predicted values
}
\description{
Plot observed and predicted catch values
}

