% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot cpue or other indices}
\usage{
plot_cpue(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of observed and predicted incices
}
\description{
Plot cpue or other indices
}

