% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_cpue_res}
\alias{plot_cpue_res}
\title{Plot residuals of cpue or other indices}
\usage{
plot_cpue_res(M)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of fit indices residuals
}
\description{
Plot residuals of cpue or other indices
}

