% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_multiple.R
\name{plot_multiple}
\alias{plot_multiple}
\title{Plot multiple figures on a page}
\usage{
plot_multiple(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{ggplot objects}

\item{file}{does nothing (yet)}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
page with figures
}
\description{
From http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\details{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}

