% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity}
\usage{
plot_selectivity(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of selectivity
}
\description{
Plot selectivity
}

