% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_sizecomp}
\alias{plot_sizecomp}
\title{Plot observed and predicted size composition}
\usage{
plot_sizecomp(replist, which_plots = "all")
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of observed and predicted size composition
}
\description{
TODO: Insert more information here.
}

