% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_sizecomp_res}
\alias{plot_sizecomp_res}
\title{Plot size composition residuals}
\usage{
plot_sizecomp_res(replist, which_plots = "all")
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Bubble plot of size composition residuals
}
\description{
TODO: Insert more information here.
}

