% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-sizetransition.R
\name{plot_sizetransition}
\alias{plot_sizetransition}
\title{Plot size transition}
\usage{
plot_sizetransition(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of size transition matrix
}
\description{
Plot size transition
}

