% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{get_selectivity}
\alias{get_selectivity}
\title{Get selectivity}
\usage{
get_selectivity(replist, type = 1)
}
\arguments{
\item{replist}{List object created by read_admb function}

\item{type}{=1 Capture, =2 Retained, =3 Discarded selectivity}
}
\value{
List of selectivities
}
\description{
Get selectivity
}

