% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot cpue or other indices}
\usage{
plot_cpue(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of all observed and predicted incices
}
\description{
Plot cpue or other indices
}

