% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-naturalmortality.R
\name{plot_naturalmortality}
\alias{plot_naturalmortality}
\title{Plot natural mortality}
\usage{
plot_naturalmortality(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot natural mortality over time and size
}
\description{
Plot natural mortality
}

