% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-sizecomp.R
\name{plot_sizecomp}
\alias{plot_sizecomp}
\title{Plot observed and predicted size composition}
\usage{
plot_sizecomp(replist, which_plots = "all", main = "")
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of observed and predicted size composition
}
\description{
TODO: Insert more information here.
}

